Fix 32 bits compile error

+++ iModelBridgeCore/ScalableMesh/PublicAPI/ScalableMesh/ScalableMeshDefs.h
@@ -168,11 +168,11 @@ enum class SMTextureType
 
 struct SMRasterTile
     {
-    size_t m_posX;          //In pixels
-    size_t m_posY;          //In pixels
-    size_t m_sizeX;         //In pixels
-    size_t m_sizeY;         //In pixels
-    size_t m_resolutionInd; 
+    uint64_t m_posX;          //In pixels
+    uint64_t m_posY;          //In pixels
+    uint64_t m_sizeX;         //In pixels
+    uint64_t m_sizeY;         //In pixels
+    uint64_t m_resolutionInd;
     };
     
 #define MEAN_SCREEN_PIXELS_PER_POINT 100
